fix: refactor county search name
