+++ UnitTests/legallead.search.tests/util/DallasActionContainerTests.cs
@@ -1,4 +1,5 @@
-﻿using LegalLead.PublicData.Search.Util;
+﻿using LegalLead.PublicData.Search.Interfaces;
+using LegalLead.PublicData.Search.Util;
 using System;
 using System.Linq;
 using Thompson.RecordSearch.Utility.Interfaces;
