+++ UnitTests/legallead.search.tests/util/DallasUiFetchTests.cs
@@ -1,5 +1,6 @@
 ﻿using Bogus;
 using LegalLead.PublicData.Search.Classes;
+using LegalLead.PublicData.Search.Interfaces;
 using LegalLead.PublicData.Search.Util;
 using Moq;
 using Newtonsoft.Json;
